+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Adding the `parsers/brown` namespace.
 * Adding the `parsers/conll` namespace.
 * Adding the `structures/bk-tree` namespace.
+* Adding the `structures/bloom-filter` namespace.
 * Adding the `sampler` options to `clustering/k-means`.
 * Adding the `stats/descriptive#.quantile` function.
 * Adding the `stats/descriptive#.median` function.